requirements.txt
streamlit
PyPDF2
pdfminer
pdfminer.six
langchain
python-dotenv
openai
chromadb
tiktoken